Implemented ipcKeyCommand for object drop
